+++ src/Bundle/AutoMapperBundle/Tests/Fixtures/Cat.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Jane\Bundle\AutoMapperBundle\Tests\Fixtures;
-
-class Cat extends Pet
-{
-}