+++ src/Component/AutoMapper/Tests/Fixtures/CityFoo.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Jane\Component\AutoMapper\Tests\Fixtures;
-
-class CityFoo
-{
-    public $name;
-}